Add test_connectivity.rs
